+++ src/history/__tests__/filterButton.test.tsx
@@ -1,11 +1,20 @@
 import { shallow } from 'enzyme'
-import FilterButton from 'history/filterButton'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { mockFilterButtonProps } from './__mock__/mockData'
+import FilterButton from 'history/filterButton'
+
+import { filters, filterType } from '../historyConst'
+import { FilterButtonProps } from '../types'
 
 describe('filterButton component', () => {
+  const mockFilterButtonProps: FilterButtonProps = {
+    type: filterType.PRICE,
+    value: filters.OPEN,
+    selected: false,
+    onClick: jest.fn()
+  }
+
   it('should render unselected filter button correctly', () => {
     const tree = renderer.create(<FilterButton {...mockFilterButtonProps} />).toJSON()
 
@@ -20,8 +29,8 @@ describe('filterButton component', () => {
 
   it('should call onClick method when button is clicked', () => {
     const wrapper = shallow(<FilterButton {...mockFilterButtonProps} />)
-
     wrapper.find('button').simulate('click')
+
     expect(mockFilterButtonProps.onClick).toHaveBeenCalledTimes(1)
   })
 })
